add extra info for feature/structure crashes
It seems to work well except that the newlines broke in Not Enough Crashes when printing to the logs which has me baffled. But I'll leave it to you to change the formatting to how you desire it in the logs.

To test, downlaod this datapack and put it in the world's datapack folder. Re-enter world and attempt to enter the nether to get a configuredfeature crash
https://github.com/TelepathicGrunt/Blame/releases/tag/0.0.0